+++ src/main/scala/hydrozoa/l2/consensus/Consensus.scala
@@ -1,1 +1,124 @@
 package hydrozoa.l2.consensus
+
+import com.typesafe.scalalogging.Logger
+import hydrozoa.l2.consensus.network.Req
+import hydrozoa.l2.consensus.network.actor.{ConsensusActor, ConsensusActorFactory}
+import hydrozoa.l2.consensus.network.transport.{AnyMsg, HeadPeerNetworkTransportWS}
+import hydrozoa.node.TestPeer
+import ox.Ox
+import ox.channels.{Actor, ActorRef, Channel, Source}
+import ox.flow.Flow
+
+import scala.collection.mutable
+
+/** Manages consensus dynamic actors and handles incoming messages.
+  */
+trait ConsensusDispatcher:
+    //
+    def setTransport(transport: ActorRef[HeadPeerNetworkTransportWS]): Unit
+    def setConsensusActorFactory(consensusActorFactory: ConsensusActorFactory): Unit
+    //
+    def dispatchMessage(anyMsg: AnyMsg): Unit
+    def spawnActorProactively(from: TestPeer, seq: Long, req: Req): Source[req.resultType]
+    def run()(using Ox): Unit
+end ConsensusDispatcher
+
+class DefaultConsensusDispatcher extends ConsensusDispatcher:
+
+    private var transport: ActorRef[HeadPeerNetworkTransportWS] = _
+
+    override def setTransport(transport: ActorRef[HeadPeerNetworkTransportWS]): Unit =
+        this.transport = transport
+
+    private var consensusActorFactory: ConsensusActorFactory = _
+
+    override def setConsensusActorFactory(consensusActorFactory: ConsensusActorFactory): Unit =
+        this.consensusActorFactory = consensusActorFactory
+
+    private val log = Logger(getClass)
+
+    private val actors: mutable.Map[(TestPeer, Long), ActorRef[ConsensusActor]] = mutable.Map.empty
+
+    def dispatchMessage(msg: AnyMsg): Unit =
+        log.info(s"Dispatching incoming message: $msg")
+
+        val origin = msg.origin
+        actors.get(origin) match
+            case Some(actor) =>
+                log.info(s"Actor was found for origin: $origin")
+                msg.asReqOrAck match
+                    case Left(originPeer, originSeq, req) =>
+                        // FIXME: add check whether init has not been called
+                        val acks = actor.ask(act => act.init(req.asInstanceOf[act.ReqType]))
+                        log.info(s"Replying with acks: $acks")
+                        acks.foreach(ack =>
+                            transport.tell(_.broadcastAck(originPeer, originSeq)(ack))
+                        )
+                    case Right(_, _, originPeer, originSeq, ack) =>
+                        val mbAck = actor.ask(act => act.deliver(ack.asInstanceOf[act.AckType]))
+                        log.info(s"Replying with mbAck: $mbAck")
+                        mbAck.foreach(ack =>
+                            transport.tell(_.broadcastAck(originPeer, originSeq)(ack))
+                        )
+            case None =>
+                log.info(s"Actor was NOT found for origin: $origin")
+                // TODO: First reaction: here would be nice to check, that if at least one
+                //  block-related consensus actor is already here we must indicate an
+                //  erroneous condition: under normal operation there should be exactly
+                //  one block in work.
+                //  After some thinking: no, we should not, since if the next leader got all
+                //  confirmations and start working on the next block theoretically we can
+                //  get a message about the next block before we finish with the previous one.
+                //  This situation should be definitely tested in simulation.
+                val mbNewActor = msg.asReqOrAck match
+                    case Left(originPeer, originSeq, req) =>
+                        val (newActor, acks) = consensusActorFactory.spawnByReq(req)
+                        acks.foreach(ack =>
+                            transport.tell(_.broadcastAck(originPeer, originSeq)(ack))
+                        )
+                        Some(newActor)
+                    case Right(_, _, originPeer, originSeq, ack) =>
+                        val mbActor -> mbAck = consensusActorFactory.spawnByAck(ack)
+                        mbAck.foreach(ack =>
+                            transport.tell(_.broadcastAck(originPeer, originSeq)(ack))
+                        )
+                        mbActor
+                mbNewActor match
+                    case Some(newActor) =>
+                        spawnActorReactivelyIn.send(newActor)
+                        val newActorRef = spawnActorReactivelyOut.receive()
+                        log.info(s"Adding new actor for $origin")
+                        actors.put(origin, newActorRef)
+                    case None => ()
+
+    override def spawnActorProactively(
+        from: TestPeer,
+        seq: Long,
+        req: Req
+    ): Source[req.resultType] =
+        val origin = (from, seq)
+        log.info("Spawning actor proactively...")
+        val (newActor, acks) = consensusActorFactory.spawnByReq(req)
+        spawnActorReactivelyIn.send(newActor)
+        val newActorRef = spawnActorReactivelyOut.receive()
+        actors.put(origin, newActorRef)
+        transport.tell(_.broadcastReq(Some(seq))(req))
+        acks.foreach(ack => transport.tell(_.broadcastAck(from, seq)(ack)))
+        log.info("Getting result source...")
+        val source: Source[req.resultType] = newActorRef.ask(act => act.result(using req))
+        source
+
+    private val spawnActorReactivelyIn: Channel[ConsensusActor] = Channel.rendezvous
+    private val spawnActorReactivelyOut: Channel[ActorRef[ConsensusActor]] = Channel.rendezvous
+
+    def run()(using Ox): Unit =
+        log.info("running reactive spawner...")
+        Flow.fromSource(spawnActorReactivelyIn)
+            .runForeach(actor =>
+                log.info(s"reactively spanning: ${actor.getClass}")
+                val actorRef = Actor.create(actor)
+                spawnActorReactivelyOut.send(actorRef)
+            )
+
+object DefaultConsensusDispatcher:
+    def apply() = new DefaultConsensusDispatcher
