+++ src/main/scala/hydrozoa/l2/consensus/network/HeadPeerNetwork.scala
@@ -5,7 +5,7 @@ import com.github.plokhotnyuk.jsoniter_scala.macros.JsonCodecMaker
 import hydrozoa.*
 import hydrozoa.l1.multisig.tx.PostDatedRefundTx
 import hydrozoa.l2.block.{Block, BlockBody, BlockHeader, BlockTypeL2}
-import hydrozoa.l2.consensus.network.transport.IncomingDispatcher
+import hydrozoa.l2.consensus.ConsensusDispatcher
 import hydrozoa.l2.ledger.*
 import hydrozoa.l2.ledger.event.NonGenesisL2EventLabel
 import hydrozoa.node.TestPeer
@@ -18,7 +18,7 @@ import scala.collection.mutable
 // FIXME: revise return types?
 trait HeadPeerNetwork {
 
-    def setDispatcher(dispatcherRef: ActorRef[IncomingDispatcher]): Unit
+    def setDispatcher(dispatcherRef: ActorRef[ConsensusDispatcher]): Unit
 
     /** @return
       *   set iof verification keys for all known peers
